Undo shared HelsinkiProfileApiTokenAuthentication
