+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 fail_fast: true
 
 default_language_version:
-    python: python3
+  python: python3
 
 repos:
   - repo: https://github.com/psf/black
@@ -9,6 +9,11 @@ repos:
     hooks:
       - id: black
         args: [--preview]
+        files: ^backend/
+  - repo: https://github.com/pycqa/isort
+    rev: 5.12.0
+    hooks:
+      - id: isort
         files: ^backend/
   - repo: https://github.com/pycqa/flake8
     rev: 4.0.1
@@ -18,8 +23,3 @@ repos:
         exclude: migrations|snapshots
         files: ^backend/
         args: [--config, backend/shared/precommit-setup.cfg]
-  - repo: https://github.com/pycqa/isort
-    rev: 5.12.0
-    hooks:
-      - id: isort
-        files: ^backend/
